backend/requirements.txt
Flask
flask_cors
psycopg2-binary
python-dotenv # For loading environment variables from a .env file
gunicorn # For production-ready WSGI HTTP server
